end exercise 4.21 cambiar la operacion de eliminacion

+++ part4Exercise/controllers/BlogController.js
@@ -18,9 +18,6 @@ blogsRouter.post('/', async (request, response, next) => {
 
   // const user = await User.findOne({})
   const decodedToken = jwt.verify(request.token, process.env.SECRET)
-  if (!request.token || !decodedToken.id) {
-    return response.status(401).json({ error: 'token missing or invalid' })
-  }
   const user = await User.findById(decodedToken.id)
 
   const blog = new Blog({
@@ -44,9 +41,30 @@ blogsRouter.post('/', async (request, response, next) => {
 
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', async (request, response, next) => {
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  const user = await User.findById(decodedToken.id)
+  const blogToDelete = await Blog.findById(request.params.id)
+
+  console.log('DELETE blogToDelete', blogToDelete)
+  console.log('DELETE user', user)
+
+  if (!blogToDelete) return response.status(204).end()
+  if (!user) return response.status(401).json({ error: 'token user not found' })
+
+
+  if (blogToDelete.user._id.toString() === user._id.toString()) {
+    try {
+      await Blog.findByIdAndRemove(request.params.id)
+      response.status(204).end()
+    } catch (exception) {
+      next(exception)
+    }
+  } else {
+    return response.status(401).json({ error: 'Unauthorized' })
+  }
+
 })
 
 blogsRouter.put('/:id', async (request, response) => {
